+++ src/code/NSError+SoapFault.m
@@ -1,0 +1,14 @@
+//
+//  NSError+SoapFault.m
+//  dXml
+//
+//  Created by Derek Clarkson on 18/01/10.
+//  Copyright 2010 Derek Clarkson. All rights reserved.
+//
+
+#import "NSError+SoapFault.h"
+
+
+@implementation NSError_SoapFault
+
+@end