+++ src/code/SoapWebServiceConnection.h
@@ -26,6 +26,7 @@ XmlParserErrorCode;
 
 #define SOAP_WEB_SERVICE_CONNECTION_DOMAIN @"dXml:SoapWebServiceConnection"
 
+
 /**
  * Extension of the UrlConnection class which is specially designed to handle soap web service calls to servers.
  * This class is the main class of the dXml API.